Remove graph functionality
