+++ tests/test_llm_summary.py
@@ -1,6 +1,5 @@
 import json, base64
 import dashboard_app as da
-go = da.go
 from scripts.judge_utils import merge_judge_results
 
 def test_summarize_judge_results():
@@ -32,23 +31,6 @@ def test_judge_timeline_trace():
     assert timeline == [1, 0]
     assert jtimeline == [1, 0]
 
-
-def test_flag_comparison_figure():
-    features = [
-        {"index": 0, "sender": "bot", "timestamp": None, "text": "buy", "flags": {"urgency": True, "emotion_count": 1}},
-        {"index": 1, "sender": "bot", "timestamp": None, "text": "please", "flags": {"guilt": True}},
-    ]
-    judge = {
-        "flagged": [
-            {"index": 0, "text": "buy", "flags": {"urgency": True}},
-            {"index": 1, "text": "please", "flags": {"guilt": True}},
-        ]
-    }
-    heur, llm = da.compute_flag_counts(features, judge)
-    assert heur["urgency"] == 1 and heur["guilt"] == 1
-    assert llm["urgency"] == 1 and llm["guilt"] == 1
-    fig = da.build_flag_comparison_figure(features, judge, "#000", "white")
-    assert fig is not None
 
 
 def test_flag_helpers_with_merged_results():
